add support for the tsize option (RFC 2349)
For RRQ, the server returns the size of the file to be transfered
in reply to the tsize option.

For WRQ, the client sends the size of the file to be transfered,
thus allowing the server to turn down some file transfers.
This patch does not add support for this behaviour.